Break ShellOut components into individual files
