+++ Sources/ShellOutError.swift
@@ -1,0 +1,46 @@
+import Foundation
+
+
+/// Error type thrown by the `shellOut()` function, in case the given command failed
+public struct ShellOutError: Swift.Error {
+    /// The termination status of the command that was run
+    public let terminationStatus: Int32
+    /// The error message as a UTF8 string, as returned through `STDERR`
+    public var message: String { return errorData.shellOutput() }
+    /// The raw error buffer data, as returned through `STDERR`
+    public let errorData: Data
+    /// The raw output buffer data, as retuned through `STDOUT`
+    public let outputData: Data
+    /// The output of the command as a UTF8 string, as returned through `STDOUT`
+    public var output: String { return outputData.shellOutput() }
+}
+
+extension ShellOutError: CustomStringConvertible {
+    public var description: String {
+        return """
+               ShellOut encountered an error
+               Status code: \(terminationStatus)
+               Message: "\(message)"
+               Output: "\(output)"
+               """
+    }
+}
+
+extension ShellOutError: LocalizedError {
+    public var errorDescription: String? {
+        return description
+    }
+}
+
+
+private extension Data {
+    func shellOutput() -> String {
+        let output = String(decoding: self, as: UTF8.self)
+
+        guard !output.hasSuffix("\n") else {
+            return String(output.dropLast())
+        }
+
+        return output
+    }
+}
